% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighting.R
\name{add_weight_columns}
\alias{add_weight_columns}
\title{Add weighting columns to a data frame for use in post-stratification}
\usage{
add_weight_columns(data)
}
\arguments{
\item{data}{a data frame in long format containing the data to recode.}
}
\value{
a version of \code{data} with new columns \code{ISCED 2011 Level} (education
  levels), and \code{NUTS2} (in which counties have been mapped to NUTS 2
  regions).
}
\description{
Add weighting columns to a data frame for use in post-stratification
}
