% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names_concepts.R
\docType{data}
\name{common_concepts}
\alias{common_concepts}
\title{Key concepts common to all four questionnaires and must be Bonferroni-corrected.}
\format{An object of class \code{character} of length 4.}
\usage{
common_concepts
}
\description{
Key concepts common to all four questionnaires and must be Bonferroni-corrected.
}
\keyword{datasets}
