% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_answers.R
\name{correct_answer_data}
\alias{correct_answer_data}
\title{Define the correct answers and key concepts in terms of quiz and question number pairs}
\usage{
correct_answer_data()
}
\description{
Define the correct answers and key concepts in terms of quiz and question number pairs
}
