% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_understanding.R
\name{data_for_concept}
\alias{data_for_concept}
\title{Obtain data on a single concept}
\usage{
data_for_concept(concept, data = quiz_data_long())
}
\arguments{
\item{concept}{the concept of interest (e.g., "Newer is not necessarily better").}

\item{data}{a data frame of class \code{quiz_long} containing the data (by default,
\code{quiz_data_long} is used).}
}
\value{
a data frame with one row per participant on the specified concept, and in
  particular a factor \code{concept_correct} that encodes whether participants
  understand the concept. Defined columns will be renamed to remove spaces from their
  names (it is assumed that objects of class \code{quiz_long} have these variables).
}
\description{
Obtain data on a single concept
}
