% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_survey.R
\name{do_get_quiz_data}
\alias{do_get_quiz_data}
\title{Read data for one of the four questionnaires}
\usage{
do_get_quiz_data(n)
}
\arguments{
\item{n}{The quiz number (1 to 4, inclusive).}
}
\value{
a \code{tibble} containing the quiz data.
}
\description{
Do not use this function directly, but prefer the (memoized) \code{get_quiz_data}.
}
