% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_understanding.R
\name{expected_probability}
\alias{expected_probability}
\title{Compute the probability that a participant would understand a concept by guessing}
\usage{
expected_probability(concept, data = quiz_data_long())
}
\arguments{
\item{concept}{the concept of interest (e.g., "Newer is not necessarily better").}

\item{data}{a data frame in long format containing the data (by default,
\code{quiz_data_long}) is used.}
}
\value{
an estimate of the proportion of Norwegians who would be expected to
  "understand" a concept if they made random guesses.
}
\description{
All possible responses to questions are considered, including "Don't know" etc.
}
