% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_norwegians_regression.R
\name{exploratory_regression_norwegians}
\alias{exploratory_regression_norwegians}
\title{Exploratory analysis of association of demographics and Norwegians' outcomes.}
\usage{
exploratory_regression_norwegians(outcome, outcome_name, family)
}
\arguments{
\item{outcome}{The outcome of interest (e.g., passing or mastery).}

\item{outcome_name}{The name of the outcome (for display in tables etc.)}

\item{family}{The error famility (and link function) to use for the outcome.}
}
\value{
an obejct of class \code{expl_reg} containg the results of the 
 estimation procedure (itself an instance of the class \code{MIresult}).
}
\description{
This function is used to perform exploratory regression analyses of the
mean score, and of passing or mastery, for Norwegians. Note that a subset
of the Norwegian respondents are used (the same as those used in the 
comparison with Ugandans). This subset is used because the repondents were
asked all the questions necessary to assess passing and mastery.
}
