% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_docs.R
\docType{package}
\name{fhi.ihc.norway.2019}
\alias{fhi.ihc.norway.2019}
\alias{fhi.ihc.norway.2019-package}
\title{A package to analyze survey data of Norwegians' understanding of IHC concepts}
\description{
A package to analyze survey data of Norwegians' understanding of IHC concepts
}
\note{
Installing this package may generate numerous messages, some of which
  are reported to be errors. These messages indicate how columns in imported
  data are being parsed; indicate parsing errors in columns not used in 
  the analyses; and note that the `exp` function is called without arguments
  when used in `magrittr` pipelines. You can safely ignore these messages.
}
