% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_num_correct_survey.R
\name{fit_survey_glm}
\alias{fit_survey_glm}
\title{Fit a GLM to a combination of the Uganda trial and Norwegian survey data}
\usage{
fit_survey_glm(data, formula, family)
}
\arguments{
\item{data}{an imputed data set with columns as named in the formula, plus
\code{country} and \code{school} columns identifying the cluster each individual
belongs to.}

\item{formula}{a formula specifying the model structure.}

\item{family}{an object that specifies the error family (and link function).}
}
\value{
an object of class \code{svyglm}.
}
\description{
The cluster structure is modelled using the method provided by the \code{survey}
package, not using random effects (which does not appear to be possible at this
time in R if post-stratified weights are to be used).
}
