% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_survey.R
\name{get_quiz_data}
\alias{get_quiz_data}
\title{Read data for one of the four questionnaires}
\usage{
get_quiz_data(n)
}
\arguments{
\item{n}{The quiz number (1 to 4, inclusive).}
}
\value{
a \code{tibble} containing the quiz data.
}
\description{
Read data for one of the four questionnaires
}
