% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_ncg_binomial.R
\name{helper_ncg_binomial}
\alias{helper_ncg_binomial}
\title{Helper for \code{summary} and \code{plot} methods.}
\usage{
helper_ncg_binomial(object)
}
\arguments{
\item{object}{an object of class \code{ncg_gaussian}.}
}
\value{
A \code{tibble} (table) of intermediate results.
}
\description{
Helper for \code{summary} and \code{plot} methods.
}
