% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_answers.R
\name{key_concept}
\alias{key_concept}
\title{The concept that a given quiz and question number pair probes.}
\usage{
key_concept(quiz_question)
}
\arguments{
\item{quiz_question}{a string like "Quiz 1 Question 2" or "Quiz 1 Question 18.1".}
}
\value{
the key concept probed. If there is no correct answer defined for the quiz and
  question number pair, the function will return "REMOVE_ME", under the assumption that
  the pair corresponds to a "false" survey participant who did not take the quiz, but
  who was created as a result of converting the original wide data frames to long
  format.
}
\description{
The concept that a given quiz and question number pair probes.
}
