% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_survey.R
\name{make_quiz_data_long}
\alias{make_quiz_data_long}
\title{Function to make data for participant responses for all quizzes, in "long" format}
\usage{
make_quiz_data_long()
}
\value{
a "long" \code{tibble} where individuals are identified by a \code{NR} column,
  and questions and participants' responses to them are provided in columns
  \code{Question} and \code{Answer}.
}
\description{
Do not call this function directly, use \code{quiz_data_long}.
}
