% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_norway_num_correct.R
\name{norwegian_num_correct}
\alias{norwegian_num_correct}
\title{Data on number of questions answered correctly for Norwegian adults}
\usage{
norwegian_num_correct()
}
\value{
a \code{tibble} where each row is a Norwegian adult, and the columns are the
  same as for the data sets \code{uganda_children_data}, \code{uganda_parents_data},
  and \code{uganda_teachers_data}.
}
\description{
The Norwegians were randomly sampled and there is no hierarchical grouping;
\code{school} is set to a unique value for each individual.
}
\details{
Because only quiz 1 was the same as in the Uganda trials, only data for that quiz is
returned. Responses on the questions that probe understanding of association and
causation are dropped from the returned data set.

The columns \code{passing} and \code{mastery} use the previously-published definitions,
but are only meaningful with respect to quiz 1.
}
