% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_num_correct_glmm.R
\name{num_correct_glmm_analysis}
\alias{num_correct_glmm_analysis}
\title{Perform a GLMM analysis of the number of correct answers of Ugandans and Norwegians.}
\usage{
num_correct_glmm_analysis(formula, family,
  reference_group = groups$intervention_children)
}
\arguments{
\item{formula}{the formula that specifies the GML analysis.}

\item{family}{the error family to use for the GLM.}

\item{reference_group}{the name of the reference group.}
}
\value{
a named list of results, of class \code{ncg_gaussian} or \code{ncg_binomial}.
}
\description{
Perform a GLMM analysis of the number of correct answers of Ugandans and Norwegians.
}
