% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_num_correct_survey.R
\name{num_correct_impute}
\alias{num_correct_impute}
\title{Set up an imputation for a given data set}
\usage{
num_correct_impute(x, exclude_cluster = TRUE)
}
\arguments{
\item{x}{a data frame with missing data.}

\item{exclude_cluster}{if \code{TRUE}, exclude cluster variables from imputation.}
}
\value{
an S3 object of class \code{mids}.
}
\description{
The function ensures that any factor variable is converted to a character variable,
permitting the various data sets to be joined without errors about differing types. We
add a probs column which by default is all set to 1; for the Norway data, this should
be rewritten manually.
}
