% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_num_correct_survey.R
\name{num_correct_impute_and_estimate}
\alias{num_correct_impute_and_estimate}
\title{Perform an analysis for an i-th completed data set.}
\usage{
num_correct_impute_and_estimate(i, analysis, uganda_teachers_data,
  uganda_children_imputed, uganda_parents_imputed, norway_imputed)
}
\arguments{
\item{i}{an index into one of the completed (imputed) data sets.}

\item{analysis}{a function of a completed data set.}

\item{uganda_teachers_data}{data for Ugandan teachers (note that this data is complete).}

\item{uganda_children_imputed}{the result of calling \code{num_correct_impute} on
\code{uganda_children_data}.}

\item{uganda_parents_imputed}{the result of calling \code{num_correct_impute} on
\code{uganda_parents_data}.}

\item{norway_imputed}{the result of calling \code{num_correct_impute} on
\code{norwegian_num_correct()}.}
}
\value{
this function returns whatever the \code{analysis} function returns.
}
\description{
Perform an analysis for an i-th completed data set.
}
