% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_num_correct_survey.R
\name{num_correct_sample_analysis}
\alias{num_correct_sample_analysis}
\title{Compute quantities to support the reporting of a GLM.}
\usage{
num_correct_sample_analysis(formula, family)
}
\arguments{
\item{formula}{the formula that specifies the GML analysis.}

\item{family}{the error family to use for the GLM.}
}
\value{
a data frame with columns to describe the sample. Note that for analyses using
  the quasibinomial error family, the columns \code{sample_n} and \code{sample_k} gives
  the number of "successes" and "failures", such that their ration gives the same odds.
}
\description{
Compute quantities to support the reporting of a GLM.
}
