% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_num_correct_survey.R
\name{num_correct_survey_analysis}
\alias{num_correct_survey_analysis}
\title{Perform an analysis of the number of correct answers of Ugandans and Norwegians.}
\usage{
num_correct_survey_analysis(formula, family)
}
\arguments{
\item{formula}{the formula that specifies the GML analysis.}

\item{family}{the error family to use for the GLM.}
}
\value{
a named list of results, which are each also given the class
  \code{num_correct_analysis}.
}
\description{
Multiple imputation is used to complete missing values of poststratification variables.
Post-stratification is used to address possible nonrandom nonresponse of Norwegians.
Clustering is modelled using the method provided by the \code{survey} package rather
than using random effects (see \code{link{fit_survey_glm}}).
}
