% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names_variables.R
\docType{data}
\name{participant_characteristics_variables}
\alias{participant_characteristics_variables}
\title{Participant characteristic variable name mappings}
\format{An object of class \code{list} of length 6.}
\usage{
participant_characteristics_variables
}
\description{
Maps from the Norwegian column names (RHS) to short English names (LHS).
}
\keyword{datasets}
