% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_participants.R
\name{participant_table_row}
\alias{participant_table_row}
\title{Helper function for making rows of table of participants}
\usage{
participant_table_row(data, filter_f)
}
\description{
Helper function for making rows of table of participants
}
