% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_participants.R
\name{participant_table_section}
\alias{participant_table_section}
\title{Helper function for making a section of a table of participants}
\usage{
participant_table_section(data, column)
}
\description{
Helper function for making a section of a table of participants
}
