% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_logit_df.R
\name{plot.logit_df}
\alias{plot.logit_df}
\title{\code{plot} method for \code{logit_df} objects.}
\usage{
\method{plot}{logit_df}(object, simple = FALSE, concept_subset = NULL)
}
\arguments{
\item{object}{an object of class \code{logit_df}.}

\item{concept_subset}{a vector of concept names to include in the plot. By default, all
concepts are included for plots of concepts (as opposed to attitudes and intentions).}
}
\value{
A \code{ggplot2} object representing the figure.
}
\description{
This method makes a forest plot to estimated odds ratios.
}
