% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_ncg_binomial.R
\name{plot.ncg_binomial}
\alias{plot.ncg_binomial}
\title{Plot a \code{ncg_binomial} object}
\usage{
\method{plot}{ncg_binomial}(object)
}
\arguments{
\item{object}{an object of class \code{ncg_gaussian}.}
}
\value{
A plot showing the results.
}
\description{
Plot a \code{ncg_binomial} object
}
