% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_ncg_gaussian.R
\name{plot.ncg_gaussian}
\alias{plot.ncg_gaussian}
\title{Plot a \code{ncg_gaussian} object}
\usage{
\method{plot}{ncg_gaussian}(object)
}
\arguments{
\item{object}{an object of class \code{ncg_gaussian}.}
}
\value{
A plot showing the results.
}
\description{
Plot a \code{ncg_gaussian} object
}
