% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_understanding.R
\name{proportion_concept_row}
\alias{proportion_concept_row}
\title{Make a one-row tibble for the proportion of Norwegians who understand a given concept}
\usage{
proportion_concept_row(concept, data = quiz_data_long())
}
\arguments{
\item{concept}{the concept (e.g., a member of \code{concepts})}

\item{data}{a data frame in long format containing the data (by default,
\code{quiz_data_long}) is used.}
}
\value{
a one-row \code{tibble} that summarizes the proportion of Norwegians who
  understand the concept. The \code{tibble} will contain a column that gives the
  half-width of the confidence interval, which may be useful for comparing with the
  original power analysis.
}
\description{
If the concept is one of the four included in all questionnaires, the confidence
interval will be Bonferroni-corrected.
}
