% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_understanding.R
\name{proportion_understand}
\alias{proportion_understand}
\title{Estimate (with/without adjustment) the proportion of Norwegians who understand a concept}
\usage{
proportion_understand(concept, data = quiz_data_long(),
  regress = FALSE)
}
\arguments{
\item{concept}{the concept of interest (e.g., "Newer is not necessarily better").}

\item{data}{a data frame in long format containing the data (by default,
\code{quiz_data_long}) is used.}

\item{regress}{if \code{FALSE} (the default), estimate the proportion of
Norwegians who understand the concept. If \code{TRUE}, use logistic
regression to model understanding the concept in terms of covariates sex,
research training, research participation, education, and medical education.}
}
\value{
an estimate of the proportion of Norwegians who understand the concept. The
  returned object will have an attribute, \code{raw_estimate}, with the raw estimate
  (i.e., without poststratification).
}
\description{
This function estimates a design-weighted mean (proportion) of respondents who
understand a given concept, or performs logistic regression to adjust for
covariates of interest and hence estimate their role in understanding the concept.
}
\details{
Multiple imputation is used to complete missing values of poststratification variables.
It is correct to perform multiple imputation for each call of this function, because
the full data set contains multiple instances of the same participant (and hence
the same pattern of missingness).
}
