% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_understanding.R
\name{proportion_understand_all_concepts}
\alias{proportion_understand_all_concepts}
\title{Estimate the proportion of Norwegians who understand each of the concepts}
\usage{
proportion_understand_all_concepts(data = quiz_data_long(),
  attitudes_intentions = FALSE)
}
\arguments{
\item{data}{a data frame in long format containing the data (by default,
\code{quiz_data_long}) is used.}

\item{attitudes_intentions}{if \code{FALSE} (the default), only questions that probe
understanding the key concepts will be included, otherwise only the questions that
probe attitudes will be included.}
}
\value{
a \code{tibble} of class \code{prop_df} that summarizes the proportion of
  Norwegians who understand the concepts. The \code{tibble} will contain a column that
  gives the half-width of the confidence interval, which may be useful for comparing
  with the original power analysis.
}
\description{
Multiple imputation is used to complete missing values of poststratification variables.
}
