% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_survey.R
\name{quiz_data_wide}
\alias{quiz_data_wide}
\title{Participants responses for all quizzes, in "wide" format}
\usage{
quiz_data_wide()
}
\value{
a "wide" \code{tibble} where each row is an individual and there is a column
  for each question.
}
\description{
Make a single data frame containing all the quiz data.
}
