% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_understanding.R
\name{regression_all_concepts}
\alias{regression_all_concepts}
\title{Exploratory regression analysis of Norwegians understanding of each of the concepts}
\usage{
regression_all_concepts(data = quiz_data_long(),
  attitudes_intentions = FALSE)
}
\arguments{
\item{data}{a data frame in long format containing the data (by default,
\code{quiz_data_long}) is used.}

\item{attitudes_intentions}{if \code{FALSE} (the default), only questions that probe
understanding the key concepts will be included, otherwise only the questions that
probe attitudes will be included.}
}
\value{
a \code{tibble} of class \code{logit_df} that summarizes the results of
  the analysis.
}
\description{
This function estimates log-ORs for an intercept term and each covariate
included in the model.
}
\details{
Multiple imputation is used to complete missing values of poststratification variables.
}
