% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_understanding.R
\name{regression_concept_row}
\alias{regression_concept_row}
\title{Make a one-row tibble for a regression analysis of Norwegians who understand a given concept}
\usage{
regression_concept_row(concept, data = quiz_data_long())
}
\arguments{
\item{concept}{the concept (e.g., a member of \code{concepts})}

\item{data}{a data frame in long format containing the data (by default,
\code{quiz_data_long}) is used.}
}
\value{
a one-row \code{tibble} that summarizes the odds ration estimated by the regression.
}
\description{
Make a one-row tibble for a regression analysis of Norwegians who understand a given concept
}
