% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_participants.R
\name{repeated_to_blank}
\alias{repeated_to_blank}
\title{Helper that changes a value to a space if the preceding value is the same}
\usage{
repeated_to_blank(x)
}
\description{
Helper that changes a value to a space if the preceding value is the same
}
