% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_expl_reg.R
\name{summary.expl_reg}
\alias{summary.expl_reg}
\title{Summarize an instance of \code{expl_reg}}
\usage{
\method{summary}{expl_reg}(object)
}
\arguments{
\item{object}{an instance of \code{expl_reg}.}
}
\value{
a 1-row \code{tibble} that summarizes the result in a readable way.
}
\description{
Summarize an instance of \code{expl_reg}
}
