% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_logit_df.R
\name{summary.logit_df}
\alias{summary.logit_df}
\title{Summarize a \code{logit_df} object as a table}
\usage{
\method{summary}{logit_df}(object)
}
\arguments{
\item{object}{an object of class \code{logit_df}.}
}
\value{
A \code{tibble} with the summary.
}
\description{
Summarize a \code{logit_df} object as a table
}
