% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_nca_gaussian.R
\name{summary.nca_gaussian}
\alias{summary.nca_gaussian}
\title{Summarize a \code{nca_gaussian} object as a table}
\usage{
\method{summary}{nca_gaussian}(object)
}
\arguments{
\item{object}{an object of class \code{nca_gaussian}.}
}
\value{
A \code{tibble} (table) of results.
}
\description{
Summarize a \code{nca_gaussian} object as a table
}
