% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_nca_quasibinomial.R
\name{summary.nca_quasibinomial}
\alias{summary.nca_quasibinomial}
\title{Summarize a \code{nca_quasibinomial} object as a table}
\usage{
\method{summary}{nca_quasibinomial}(object,
  reference_group = groups$norwegians, scale_to_sample_odds = TRUE)
}
\arguments{
\item{object}{an object of class \code{nca_quasibinomial}.}

\item{scale_to_sample_odds}{if \code{TRUE}, scale the estimated odds ratios to match
the sample odds. This is useful to address what appears to be a numerical issue (see
comments in the source for this function for details).}
}
\value{
A \code{tibble} (table) of results.
}
\description{
Summarize a \code{nca_quasibinomial} object as a table
}
