% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_ncg_binomial.R
\name{summary.ncg_binomial}
\alias{summary.ncg_binomial}
\title{Summarize a \code{ncg_binomial} object as a table}
\usage{
\method{summary}{ncg_binomial}(object)
}
\arguments{
\item{object}{an object of class \code{nca_quasibinomial}.}
}
\value{
A \code{tibble} (table) of results.
}
\description{
Summarize a \code{ncg_binomial} object as a table
}
