% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_ncg_gaussian.R
\name{summary.ncg_gaussian}
\alias{summary.ncg_gaussian}
\title{Summarize a \code{ncg_gaussian} object as a table}
\usage{
\method{summary}{ncg_gaussian}(object)
}
\arguments{
\item{object}{an object of class \code{ncg_gaussian}.}
}
\value{
A \code{tibble} (table) of results.
}
\description{
Summarize a \code{ncg_gaussian} object as a table
}
