% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_prop_df.R
\name{summary.prop_df}
\alias{summary.prop_df}
\title{Summarize a \code{prop_df} object as a table}
\usage{
\method{summary}{prop_df}(object)
}
\arguments{
\item{object}{an object of class \code{prop_df}.}
}
\value{
A \code{tibble} with the table of participant characteristics.
}
\description{
Summarize a \code{prop_df} object as a table
}
