% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_participants.R
\name{table_participants}
\alias{table_participants}
\title{Make a table of participant characteristics}
\usage{
table_participants()
}
\value{
A \code{tibble} with the table of participant characteristics.
}
\description{
Make a table of participant characteristics
}
