% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translate_education}
\alias{translate_education}
\title{Translate an Education column of a data frame from Norwegian to English}
\usage{
translate_education(data)
}
\description{
Translate an Education column of a data frame from Norwegian to English
}
