% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translate_sex}
\alias{translate_sex}
\title{Translate a Sex column of a data frame from Norwegian to English}
\usage{
translate_sex(data)
}
\description{
Translate a Sex column of a data frame from Norwegian to English
}
