% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_uganda_children.R
\docType{data}
\name{uganda_children_correct_answers}
\alias{uganda_children_correct_answers}
\title{Correct answers for Ugandan children}
\format{An object of class \code{list} of length 24.}
\usage{
uganda_children_correct_answers
}
\description{
Correct answers for Ugandan children
}
\keyword{datasets}
