% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_uganda_children.R
\docType{data}
\name{uganda_children_data}
\alias{uganda_children_data}
\title{Data for Ugandan children (one-year follow-up)}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 3943 rows and 11 columns.}
\usage{
uganda_children_data
}
\description{
The \code{ISCED_Level} column codes educational attainment using the same ranges of ISCED 2011
levels used for the Norwegian population. In this study, schools were the unit of
randomization, and a \code{school} column is provided that records which school each
student attended. Columns \code{passing} and \code{mastery} are use the
previously-published definitions.
}
\keyword{datasets}
