% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_uganda_parents.R
\docType{data}
\name{uganda_parents_data}
\alias{uganda_parents_data}
\title{Data on Ugandan control parents tested one year after the intervention}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 256 rows and 11 columns.}
\usage{
uganda_parents_data
}
\description{
The Education column codes educational attainment using the same ranges of ISCED 2011
levels used for the Norwegian population.
}
\keyword{datasets}
