% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_uganda_parents.R
\docType{data}
\name{uganda_parents_immediate_correct_answers}
\alias{uganda_parents_immediate_correct_answers}
\title{Correct answers for Ugandan parents}
\format{An object of class \code{list} of length 18.}
\usage{
uganda_parents_immediate_correct_answers
}
\description{
Correct answers for Ugandan parents
}
\keyword{datasets}
