% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_uganda_teachers.R
\docType{data}
\name{uganda_teachers_data}
\alias{uganda_teachers_data}
\title{Data for Ugandan teachers (one-year follow-up)}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 137 rows and 11 columns.}
\usage{
uganda_teachers_data
}
\description{
The \code{ISCED_Level} column codes educational attainment using the same ranges of
ISCED 2011 levels used for the Norwegian population. Columns \code{passing} and
\code{mastery} are use the previously-published definitions.
}
\keyword{datasets}
