% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighting.R
\name{weighted_design}
\alias{weighted_design}
\title{Obtain a weighted survey design object for a data set (e.g., for a concept)}
\usage{
weighted_design(data)
}
\arguments{
\item{data}{a data frame in long format containing the data.}
}
\value{
a survey design object.
}
\description{
Obtain a weighted survey design object for a data set (e.g., for a concept)
}
